Combine Cuda and SYCL test
